chore: make the maven central publishing work
Fix declaration of the central plugin
The plugin was only declared in the plugins management, but not in the plugins, so it was not actually used.
There was also a typo in the property declaring the version of the plugin

No longer push snapshots to maven central
This is not done for other connectors, this requires extra configuration that should be
shared between connector, so defer this task.

Bump maven and wrapper
Use a version of the wrapper provided by the Apache Foundation.
Do not store the jar in the repository to prevent malicious code hidden within the binary (the jar is downloaded when not on disk).

Update the build workflow to not run when only markdown files change.